Admin should be able to list apps in any space of current org
Story in CLI [#91292742] for addition of the -s flag in cf apps command.

As of now this command lists only the apps in the current space. With this new flag, user can get the list of apps in the space which is provided by the user.

Output after modification :
 apps -s spaceName
Getting apps in org orgName / space spaceName as admin...
OK

name      requested state   instances   memory   disk   urls
fib-cpu   started           1/1         256M     1G     fib-cpu.10.244.0.34.xip.io